Put, post operations implemented for environments
Implementation of environment operations moved to library package.
Code for searching objects by ids and names extracted to the library
function.
Exceptions propagation added to example config. It is useful for
testing and troubleshooting to have error message not only in logs
but in the API response too.
Tuningbox errors hierarchy added to the project.

Change-Id: Ic2fd3c3c17409723bfa3cfff1c0bb18f3a65f0d7