Components operations handled in the fuel2
Cascade deletion of resource definitons on the component deletion
fixed. Cascade only on the DB level is not enougth for deletion
referenced objects.
Resources definitions creation is not supported yet in the
component creation call. Resource definition would be able to
add to the component in the create resource definition call.
Update of resource definitions list is supported in the component
creation call.
Standard cliff formatters are used for fuel2 components operations.

Change-Id: If572f7437f48bdde65de114f6f70af3c071c1d0e